feat(alwaysontop): refactor toolbox
Use the new abstractions, which already take care of the rendering part.
