+++ react/features/always-on-top/AlwaysOnTop.js
@@ -2,7 +2,7 @@
 
 import React, { Component } from 'react';
 
-import ToolboxAlwaysOnTop from './ToolboxAlwaysOnTop';
+import Toolbar from './Toolbar';
 
 const { api } = window.alwaysOnTop;
 
@@ -15,13 +15,9 @@ const TOOLBAR_TIMEOUT = 4000;
  * The type of the React {@code Component} state of {@link FeedbackButton}.
  */
 type State = {
-    audioAvailable: boolean,
-    audioMuted: boolean,
     avatarURL: string,
     displayName: string,
     isVideoDisplayed: boolean,
-    videoAvailable: boolean,
-    videoMuted: boolean,
     visible: boolean
 };
 
@@ -45,19 +41,12 @@ export default class AlwaysOnTop extends Component<*, State> {
 
         this.state = {
             visible: true,
-            audioMuted: false,
-            videoMuted: false,
-            audioAvailable: false,
-            videoAvailable: false,
             displayName: '',
             isVideoDisplayed: true,
             avatarURL: ''
         };
 
         // Bind event handlers so they are only bound once per instance.
-        this._audioAvailabilityListener
-            = this._audioAvailabilityListener.bind(this);
-        this._audioMutedListener = this._audioMutedListener.bind(this);
         this._avatarChangedListener = this._avatarChangedListener.bind(this);
         this._largeVideoChangedListener
             = this._largeVideoChangedListener.bind(this);
@@ -66,33 +55,6 @@ export default class AlwaysOnTop extends Component<*, State> {
         this._mouseMove = this._mouseMove.bind(this);
         this._onMouseOut = this._onMouseOut.bind(this);
         this._onMouseOver = this._onMouseOver.bind(this);
-        this._videoAvailabilityListener
-            = this._videoAvailabilityListener.bind(this);
-        this._videoMutedListener = this._videoMutedListener.bind(this);
-    }
-
-    _audioAvailabilityListener: ({ available: boolean }) => void;
-
-    /**
-     * Handles audio available api events.
-     *
-     * @param {{ available: boolean }} status - The new available status.
-     * @returns {void}
-     */
-    _audioAvailabilityListener({ available }) {
-        this.setState({ audioAvailable: available });
-    }
-
-    _audioMutedListener: ({ muted: boolean }) => void;
-
-    /**
-     * Handles audio muted api events.
-     *
-     * @param {{ muted: boolean }} status - The new muted status.
-     * @returns {void}
-     */
-    _audioMutedListener({ muted }) {
-        this.setState({ audioMuted: muted });
     }
 
     _avatarChangedListener: () => void;
@@ -200,8 +162,6 @@ export default class AlwaysOnTop extends Component<*, State> {
         this._hovered = true;
     }
 
-    _videoAvailabilityListener: ({ available: boolean }) => void;
-
     /**
      * Renders display name and avatar for the on stage participant.
      *
@@ -235,64 +195,17 @@ export default class AlwaysOnTop extends Component<*, State> {
     }
 
     /**
-     * Handles audio available api events.
-     *
-     * @param {{ available: boolean }} status - The new available status.
-     * @returns {void}
-     */
-    _videoAvailabilityListener({ available }) {
-        this.setState({ videoAvailable: available });
-    }
-
-    _videoMutedListener: ({ muted: boolean }) => void;
-
-    /**
-     * Handles video muted api events.
-     *
-     * @param {{ muted: boolean }} status - The new muted status.
-     * @returns {void}
-     */
-    _videoMutedListener({ muted }) {
-        this.setState({ videoMuted: muted });
-    }
-
-    /**
      * Sets mouse move listener and initial toolbar timeout.
      *
      * @inheritdoc
      * @returns {void}
      */
     componentDidMount() {
-        api.on('audioMuteStatusChanged', this._audioMutedListener);
-        api.on('videoMuteStatusChanged', this._videoMutedListener);
-        api.on('audioAvailabilityChanged', this._audioAvailabilityListener);
-        api.on('videoAvailabilityChanged', this._videoAvailabilityListener);
         api.on('largeVideoChanged', this._largeVideoChangedListener);
         api.on('displayNameChange', this._displayNameChangedListener);
         api.on('avatarChanged', this._avatarChangedListener);
 
         this._largeVideoChangedListener();
-
-        Promise.all([
-            api.isAudioMuted(),
-            api.isVideoMuted(),
-            api.isAudioAvailable(),
-            api.isVideoAvailable()
-        ])
-        .then(([
-            audioMuted = false,
-            videoMuted = false,
-            audioAvailable = false,
-            videoAvailable = false
-        ]) =>
-            this.setState({
-                audioMuted,
-                videoMuted,
-                audioAvailable,
-                videoAvailable
-            })
-        )
-        .catch(console.error);
 
         window.addEventListener('mousemove', this._mouseMove);
 
@@ -306,14 +219,6 @@ export default class AlwaysOnTop extends Component<*, State> {
      * @returns {void}
      */
     componentWillUnmount() {
-        api.removeListener('audioMuteStatusChanged',
-            this._audioMutedListener);
-        api.removeListener('videoMuteStatusChanged',
-            this._videoMutedListener);
-        api.removeListener('audioAvailabilityChanged',
-            this._audioAvailabilityListener);
-        api.removeListener('videoAvailabilityChanged',
-            this._videoAvailabilityListener);
         api.removeListener('largeVideoChanged',
             this._largeVideoChangedListener);
         api.removeListener('displayNameChange',
@@ -343,14 +248,10 @@ export default class AlwaysOnTop extends Component<*, State> {
     render() {
         return (
             <div id = 'alwaysOnTop'>
-                <ToolboxAlwaysOnTop
-                    audioAvailable = { this.state.audioAvailable }
-                    audioMuted = { this.state.audioMuted }
+                <Toolbar
                     className = { this.state.visible ? 'fadeIn' : 'fadeOut' }
                     onMouseOut = { this._onMouseOut }
-                    onMouseOver = { this._onMouseOver }
-                    videoAvailable = { this.state.videoAvailable }
-                    videoMuted = { this.state.videoMuted } />
+                    onMouseOver = { this._onMouseOver } />
                 {
                     this._renderVideoNotAvailableScreen()
                 }
